refactor: remove ingress-nginx hello-world test
The ingress-nginx-based hello-world test was gated on IngressNginxSupported,
which defaulted to false and was never enabled by any provider suite. The
test was permanently dead code.

getWorkloadClusterDnsZone() is moved into hello_gateway.go, the only
remaining caller. IngressNginxSupported is removed from TestConfig.
Associated helloworld_values.yaml and ingress-nginx_values.yaml test data
files are deleted from all provider directories.
