--- providers/capa/china/test_data/helloworld_values.yaml
@@ -1,13 +1,0 @@
-ingress:
-  annotations:
-    kubernetes.io/tls-acme: "true"
-    cert-manager.io/cluster-issuer: letsencrypt-giantswarm
-  hosts:
-    - host: {{ index .ExtraValues "IngressUrl" }}
-      paths:
-        - path: /
-          pathType: ImplementationSpecific
-  tls:
-    - secretName: hello-world-tls
-      hosts:
-        - {{ index .ExtraValues "IngressUrl" }}
