--- providers/capa/private/test_data/ingress-nginx_values.yaml
@@ -1,11 +1,0 @@
-controller:
-  service:
-    enabled: true
-    external:
-      enabled: false
-    internal:
-      enabled: true
-      # We need to reuse the external subdomain for the internal service because dns-operator-route53 creates
-      # a wildcard record pointing towards ingress.<blah> domain.
-      subdomain: ingress
-
